requirements.txt
playwright==1.54
python-telegram-bot>=21.0
SQLAlchemy>=2.0
aiomysql